+++ hooks/useEventRelatedEvents.ts
@@ -1,7 +1,7 @@
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { Event } from "nostr-tools";
 import { nostrQueryKeys, useNostrEvents } from "@/providers/NostrEventProvider";
-import { useCallback } from "react";
+import { useCallback, useMemo } from "react";
 import { useAuth } from "./useAuth";
 import {
   parseZapRequestFromReceipt,
@@ -13,7 +13,8 @@ import {
 
 interface UseEventRelatedEvents {
   reactions: Event[];
-  replies: Event[];
+  topLevelReplies: Event[];
+  getChildReplies: (parentId: string) => Event[];
   reposts: Event[];
   genericReposts: Event[];
   zapReceipts: Event[];
@@ -26,11 +27,22 @@ interface UseEventRelatedEvents {
   refetch: () => void;
 }
 
+// Helper functions for reply hierarchy
+const hasRootTag = (reply: Event, commentId: string): boolean =>
+  reply.tags.some((tag) => tag.includes("root") && tag.includes(commentId));
+
+const hasNonRootReplyTag = (reply: Event, commentId: string): boolean =>
+  reply.tags.some((tag) => tag.includes("reply") && !tag.includes(commentId));
+
+const isRootReply = (reply: Event, commentId: string): boolean =>
+  hasRootTag(reply, commentId) && !hasNonRootReplyTag(reply, commentId);
+
 export const useEventRelatedEvents = (event: Event): UseEventRelatedEvents => {
   const { getEventRelatedEvents } = useNostrEvents();
   const queryClient = useQueryClient();
   const queryKey = nostrQueryKeys.eventRelatedEvents(event.id);
   const { pubkey } = useAuth();
+
   const {
     data: eventsCache = {},
     isLoading,
@@ -59,10 +71,22 @@ export const useEventRelatedEvents = (event: Event): UseEventRelatedEvents => {
   const { replies, reactions, reposts, genericReposts, zapReceipts } =
     getRelatedEventsFromCache(eventsCache);
 
+  const topLevelReplies = replies.filter((reply) =>
+    isRootReply(reply, event.id),
+  );
+
+  const getChildReplies = (parentId: string): Event[] =>
+    replies.filter(
+      (reply) =>
+        hasRootTag(reply, event.id) &&
+        reply.tags.some(
+          (tag) => tag.includes("reply") && tag.includes(parentId),
+        ),
+    );
+
   const userHasReacted = reactions.some((e) => e.pubkey === pubkey);
   const userHasZapped = zapReceipts.reduce((acc, e) => {
     const { receipt, amount } = parseZapRequestFromReceipt(e);
-
     return !!amount && receipt?.pubkey === pubkey ? acc + amount : acc;
   }, 0);
 
@@ -72,9 +96,10 @@ export const useEventRelatedEvents = (event: Event): UseEventRelatedEvents => {
   }, 0);
 
   return {
-    replies,
     reactions,
     reposts,
+    topLevelReplies,
+    getChildReplies,
     genericReposts,
     zapReceipts,
     zapTotal,
